Reduced SideBar Menu
Shifted Organizations and Business Info into one tab in setting page.

2 tabs in setting page --> Basic and Advanced Settings

Basic Setting --> Setting options from before
Advanced Setting --> Business Info and Organizations pages combined.
